requirements.txt
# Django Core
Django==4.2.17
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
drf-spectacular==0.27.2

# CORS
django-cors-headers==4.6.0

# Validación y formularios
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

# Gestión de archivos
Pillow==11.0.0
django-storages==1.14.4

# Utilidades
python-dotenv==1.0.1
pytz==2024.2

# Servidor WSGI
gunicorn==23.0.0
whitenoise==6.8.2

# Monitoreo y logs
django-extensions==3.2.3

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==30.8.2

# Code quality
flake8==7.1.1
black==24.10.0
isort==5.13.2

# Seguridad
django-ratelimit==4.1.0
